integrate tests with runtests
