--- test-writer-compat/Project.toml
@@ -1,5 +1,0 @@
-[deps]
-LibArchive_jll = "1e303b3e-d4db-56ce-88c4-91e52606a1a8"
-PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
-ZipFile = "a5390f91-8eb1-5f08-bee0-b1d1ffed6cea"
-p7zip_jll = "3f19e933-33d8-53b3-aaab-bd5110c3b7a0"
